+++ src/steps/storage/constants.ts
@@ -2,3 +2,5 @@ export const CLOUD_STORAGE_BUCKET_ENTITY_CLASS = 'DataStore';
 export const CLOUD_STORAGE_BUCKET_ENTITY_TYPE = 'google_storage_bucket';
 
 export const STEP_CLOUD_STORAGE_BUCKETS = 'fetch-cloud-storage-buckets';
+export const STEP_CLOUD_STORAGE_BUCKETS_WITH_ACCESS_LEVELS =
+  'fetch-cloud-storage-buckets-with-access-levels';