+++ src/steps/storage/converters.test.ts
@@ -8,7 +8,7 @@ describe('#createCloudStorageBucketEntity', () => {
       createCloudStorageBucketEntity({
         data: getMockStorageBucket(),
         projectId: DEFAULT_INTEGRATION_CONFIG_PROJECT_ID,
-        isPublic: false,
+        accessLevel: 'private',
       }),
     ).toMatchSnapshot();
   });
@@ -18,7 +18,7 @@ describe('#createCloudStorageBucketEntity', () => {
       createCloudStorageBucketEntity({
         data: getMockStorageBucket(),
         projectId: DEFAULT_INTEGRATION_CONFIG_PROJECT_ID,
-        isPublic: true,
+        accessLevel: 'public',
       }),
     ).toMatchSnapshot();
   });
@@ -37,7 +37,7 @@ describe('#createCloudStorageBucketEntity', () => {
           },
         }),
         projectId: DEFAULT_INTEGRATION_CONFIG_PROJECT_ID,
-        isPublic: false,
+        accessLevel: 'private',
       }),
     ).toMatchSnapshot();
   });
@@ -53,7 +53,7 @@ describe('#createCloudStorageBucketEntity', () => {
           },
         }),
         projectId: DEFAULT_INTEGRATION_CONFIG_PROJECT_ID,
-        isPublic: false,
+        accessLevel: 'private',
       }),
     ).toMatchSnapshot();
   });
@@ -65,20 +65,79 @@ describe('#createCloudStorageBucketEntity', () => {
           iamConfiguration: undefined,
         }),
         projectId: DEFAULT_INTEGRATION_CONFIG_PROJECT_ID,
-        isPublic: false,
+        accessLevel: 'private',
       }),
     ).toMatchSnapshot();
   });
 
-  test('should set "public" to "false" if "isPublic" is "undefined"', () => {
+  test('should set "public" to "false" if "accessLevel" is "undefined"', () => {
     expect(
       createCloudStorageBucketEntity({
         data: getMockStorageBucket({
           iamConfiguration: undefined,
         }),
         projectId: DEFAULT_INTEGRATION_CONFIG_PROJECT_ID,
-        isPublic: undefined,
+        accessLevel: undefined,
       }),
     ).toMatchSnapshot();
+  });
+
+  test('should set "accessLevel" to "SubjectToObjectACLs" when "uniformBucketLevelAccess" is not enabled and accessLevel is "private"', () => {
+    expect(
+      createCloudStorageBucketEntity({
+        data: getMockStorageBucket({
+          iamConfiguration: {
+            uniformBucketLevelAccess: {
+              enabled: false,
+            },
+          },
+        }),
+        projectId: DEFAULT_INTEGRATION_CONFIG_PROJECT_ID,
+        accessLevel: 'private',
+      }),
+    ).toMatchObject({
+      accessLevel: 'SubjectToObjectACLs',
+      accessLevelIsConditional: true,
+      accessLevelCondition: 'Subject to object ACLs',
+    });
+  });
+
+  test('should set "accessLevel" to whatever the access level is when "uniformBucketLevelAccess" is not enabled and accessLevel is not "private"', () => {
+    expect(
+      createCloudStorageBucketEntity({
+        data: getMockStorageBucket({
+          iamConfiguration: {
+            uniformBucketLevelAccess: {
+              enabled: false,
+            },
+          },
+        }),
+        projectId: DEFAULT_INTEGRATION_CONFIG_PROJECT_ID,
+        accessLevel: 'publicRead',
+        condition: 'condition',
+      }),
+    ).toMatchObject({
+      accessLevel: 'publicRead',
+      accessLevelIsConditional: true,
+      accessLevelCondition: 'condition',
+    });
+
+    expect(
+      createCloudStorageBucketEntity({
+        data: getMockStorageBucket({
+          iamConfiguration: {
+            uniformBucketLevelAccess: {
+              enabled: false,
+            },
+          },
+        }),
+        projectId: DEFAULT_INTEGRATION_CONFIG_PROJECT_ID,
+        accessLevel: 'publicWrite',
+      }),
+    ).toMatchObject({
+      accessLevel: 'publicWrite',
+      accessLevelIsConditional: false,
+      accessLevelCondition: undefined,
+    });
   });
 });