+++ src/steps/storage/converters.ts
@@ -5,6 +5,7 @@ import {
   CLOUD_STORAGE_BUCKET_ENTITY_CLASS,
 } from './constants';
 import { createGoogleCloudIntegrationEntity } from '../../utils/entity';
+import { AccessData } from '../../utils/jobState';
 
 export function getCloudStorageBucketWebLink(
   data: storage_v1.Schema$Bucket,
@@ -20,12 +21,15 @@ export function getCloudStorageBucketKey(id: string) {
 export function createCloudStorageBucketEntity({
   data,
   projectId,
-  isPublic,
+  accessLevel,
+  condition,
 }: {
   data: storage_v1.Schema$Bucket;
   projectId: string;
-  isPublic?: boolean;
-}) {
+} & AccessData) {
+  const isSubjectToObjectAcls =
+    data.iamConfiguration?.uniformBucketLevelAccess?.enabled !== true &&
+    accessLevel === 'private';
   return createGoogleCloudIntegrationEntity(data, {
     entityData: {
       source: data,
@@ -51,7 +55,24 @@ export function createCloudStorageBucketEntity({
         retentionPolicyEnabled: data.retentionPolicy?.isLocked,
         retentionPeriod: data.retentionPolicy?.retentionPeriod,
         retentionDate: data.retentionPolicy?.effectiveTime,
-        public: isPublic || false,
+        /**
+         * @deprecated - use `accessLevel` instead.
+         * `public` can be incorrect when `uniformBucketLevelAccess` is not enabled. If not enabled,
+         * each item in the google_storage_bucket is subject to its own ACL which may or may not be
+         * open to the internet. Ingesting the ACLs of every element in a storage bucket is far too
+         * large a task for this integration. This key is maintaining exclusively for backwards
+         * compatibility.
+         */
+        public: !!accessLevel && accessLevel !== 'private',
+        // It is not possible to know if bucket is public or not when uniformBucketLevelAccess is not enabled
+        // https://cloud.google.com/storage/docs/cloud-console?&_ga=2.84754521.-1526178294.1622832983&_gac=1.262728446.1626996208.CjwKCAjwruSHBhAtEiwA_qCppsTtaBT90RDQ-e9xjNnNQM0lwd2aI9wJfUhrVgFjQ0_SDu4kR1yUDhoCeRwQAvD_BwE#_sharingdata
+        accessLevel: isSubjectToObjectAcls
+          ? 'SubjectToObjectACLs'
+          : accessLevel,
+        accessLevelIsConditional: isSubjectToObjectAcls ? true : !!condition,
+        accessLevelCondition: isSubjectToObjectAcls
+          ? 'Subject to object ACLs'
+          : condition,
         // Rely on the value of the classification tag
         classification: null,
         etag: data.etag,