+++ CHANGELOG.md
@@ -8,11 +8,24 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
+- Added support for ingesting the following **new** properties on
+  `google_storage_bucket`:
+
+  | Property                   | Type                                                                    | Description                                           |
+  | -------------------------- | ----------------------------------------------------------------------- | ----------------------------------------------------- |
+  | `accessLevel`              | 'private', 'public', 'publicRead', 'publicWrite', 'subjectToObjectACLs' | The level public access on the contents of the bucket |
+  | `accessLevelIsConditional` | boolean                                                                 | Whether there is a condition to the access or not     |
+  | `accessLevelCondition`     | string                                                                  | The condition for access at this access level         |
+
 ### Changed
 
-- Changed `google_iam_bindings.permissions: string[]` to `google_iam_bindings.permissions: string` due to
-  limitations on `string[]` property lengths in JupiterOne. Queries will need to be changed from
-  `permissions={{string}}` to `permissions~={{string}}` to maintain functionality.
+- Changed `google_iam_bindings.permissions: string[]` to
+  `google_iam_bindings.permissions: string` due to limitations on `string[]`
+  property lengths in JupiterOne. Queries will need to be changed from
+  `permissions={{string}}` to `permissions~={{string}}` to maintain
+  functionality.
 
 ## 0.47.4